+++ tests/test_inject.py
@@ -47,19 +47,9 @@ def test_inject_include_apps(pipx_temp_env, capsys, with_suffix):
 
     if suffix:
         assert run_pipx_cli(
-            [
-                "inject",
-                "pycowsay",
-                PKG["black"]["spec"],
-                "--include-deps",
-            ]
+            ["inject", "pycowsay", PKG["black"]["spec"], "--include-deps"]
         )
 
     assert not run_pipx_cli(
-        [
-            "inject",
-            f"pycowsay{suffix}",
-            PKG["black"]["spec"],
-            "--include-deps",
-        ]
+        ["inject", f"pycowsay{suffix}", PKG["black"]["spec"], "--include-deps"]
     )